+++ resources/scripts/api/setLocalePreference.ts
@@ -1,9 +1,0 @@
-import http from '@/api/http';
-
-export default (locale: string): Promise<void> => {
-    return new Promise((resolve, reject) => {
-        http.post('/locales/locale', { locale })
-            .then(() => resolve())
-            .catch(reject);
-    });
-};