+++ resources/scripts/components/dashboard/forms/UpdateLanguageForm.tsx
@@ -1,12 +1,11 @@
 import React, { useState } from 'react';
-import { Actions, State, useStoreActions, useStoreState } from 'easy-peasy';
-import SpinnerOverlay from '@/components/elements/SpinnerOverlay';
+import { State, useStoreState } from 'easy-peasy';
 import { ApplicationStore } from '@/state';
 import Label from '@/components/elements/Label';
 import Select from '@/components/elements/Select';
 import tw from 'twin.macro';
 import { Button } from '@/components/elements/button/index';
-import { httpErrorToHuman } from '@/api/http';
+import { useLocation } from 'react-router-dom';
 import { t } from '@/lib/locale';
 
 export default () => {
@@ -14,10 +13,8 @@ export default () => {
     const availableLanguages = useStoreState(
         (state: State<ApplicationStore>) => state.settings.data?.availableLanguages ?? {}
     );
-    const updateLanguage = useStoreActions((actions: Actions<ApplicationStore>) => actions.user.updateUserLanguage);
-    const { clearFlashes, addFlash } = useStoreActions((actions: Actions<ApplicationStore>) => actions.flashes);
     const [language, setLanguage] = useState(user?.language ?? 'en');
-    const [isSubmitting, setIsSubmitting] = useState(false);
+    const location = useLocation();
 
     const hasChanges = language !== user?.language;
 
@@ -32,39 +29,25 @@ export default () => {
             return;
         }
 
-        clearFlashes('account:language');
-        setIsSubmitting(true);
-
-        updateLanguage(language)
-            .then(() => window.location.reload())
-            .catch((error) => {
-                setIsSubmitting(false);
-                addFlash({
-                    type: 'error',
-                    key: 'account:language',
-                    title: t('ui.common.error'),
-                    message: httpErrorToHuman(error),
-                });
-            });
+        const params = new URLSearchParams(location.search);
+        params.set('locale', language);
+        window.location.assign(`${location.pathname}?${params.toString()}`);
     };
 
     return (
-        <React.Fragment>
-            <SpinnerOverlay size={'large'} visible={isSubmitting} />
-            <form css={tw`m-0`} onSubmit={submit}>
-                <Label htmlFor={'language'}>{t('ui.auth.language_label')}</Label>
-                <Select id={'language'} value={language} onChange={(event) => setLanguage(event.currentTarget.value)}>
-                    {Object.entries(availableLanguages).map(([key, value]) => (
-                        <option key={key} value={key}>
-                            {value}
-                        </option>
-                    ))}
-                </Select>
-                <p className={'input-help'}>{t('ui.dashboard.language_preferences_description')}</p>
-                <div css={tw`mt-6`}>
-                    <Button disabled={isSubmitting || !hasChanges}>{t('ui.common.save')}</Button>
-                </div>
-            </form>
-        </React.Fragment>
+        <form css={tw`m-0`} onSubmit={submit}>
+            <Label htmlFor={'language'}>{t('ui.auth.language_label')}</Label>
+            <Select id={'language'} value={language} onChange={(event) => setLanguage(event.currentTarget.value)}>
+                {Object.entries(availableLanguages).map(([key, value]) => (
+                    <option key={key} value={key}>
+                        {value}
+                    </option>
+                ))}
+            </Select>
+            <p className={'input-help'}>{t('ui.dashboard.language_preferences_description')}</p>
+            <div css={tw`mt-6`}>
+                <Button disabled={!hasChanges}>{t('ui.common.save')}</Button>
+            </div>
+        </form>
     );
 };