+++ routes/api-client.php
@@ -29,7 +29,6 @@ Route::prefix('/account')->middleware(AccountSubject::class)->group(function () 
     });
 
     Route::put('/email', [Client\AccountController::class, 'updateEmail'])->name('api:client.account.update-email');
-    Route::put('/language', [Client\AccountController::class, 'updateLanguage'])->name('api:client.account.update-language');
     Route::put('/password', [Client\AccountController::class, 'updatePassword'])->name('api:client.account.update-password');
 
     Route::get('/activity', Client\ActivityLogController::class)->name('api:client.account.activity');